+++ Firestore/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+- [feature] Added `Query.count()`, which fetches the number of documents in the
+  result set without actually downloading the documents (#10246).
+
 # 10.0.0
 - [fixed] Fixed compiler warning about `@param comparator` (#10226).